feat(xcpd): express in lambda for routes
Refs: #1329
